Mobile debugging: option to auto-start Mobile Chrome & populate the URL

+++ com.github.sdbg.debug.core/src/com/github/sdbg/debug/core/internal/webkit/model/WebkitDebugTarget.java
@@ -516,7 +516,11 @@ public class WebkitDebugTarget extends WebkitDebugElement implements IBreakpoint
     } catch (IOException e) {
 
     }
-    process.terminate();
+
+    if (process != null) {
+      process.terminate();
+    }
+
     process = null;
   }
